// vouchainclient/src/app/model/employee-dto.ts
export class EmployeeDTO {
  errorDescription: string;
  status: string;
  usrId: string;
  usrBchAddress: string;
  usrPrivateKey: string;
  usrEmail: string;
  usrPassword: string;
  usrActive: string;
  empFirstName: string;
  empLastName: string;
  empMatricola: string;
  empInvitationCode: string;
  empCheckedLogin: string;
  cpyUsrId: string;
}